refactor: change file structure. Remove api from components, creates styles folder
