+++ src/api/authApi.ts
@@ -1,5 +1,5 @@
 import { baseUrl, path } from './basicApi';
-import { Isignin, Iregist } from '../../types/interface';
+import { Isignin, Iregist } from '../types/interface';
 
 export async function signinApi(body?: Isignin) {
   let params = {
