+++ src/components/textbook/textbook.ts
@@ -1,6 +1,6 @@
 import './textbook.scss';
-import { addWordsUserApi, getUserIdWords, UpdateWordsUserApi, getUserAllWords } from './../api/wordsApi';
-import { baseUrl, getWords, getWordId } from './../api/basicApi';
+import { addWordsUserApi, getUserIdWords, UpdateWordsUserApi, getUserAllWords } from '../../api/wordsApi';
+import { baseUrl, getWords, getWordId } from '../../api/basicApi';
 import { IapiRequestWords } from '../../types/interface';
 import { GetLocalStorageToken } from './../authorization/auth';
