+++ src/styles/common-style/btns.scss
