+++ src/styles/common-style/vars.scss
