+++ src/styles/section-style/benefits.scss
